requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.0.12
idna==3.4
Jinja2==3.0.3
jinja2-cli==0.8.2
lxml==4.9.2
MarkupSafe==2.0.1
PyYAML==6.0
requests==2.27.1
soupsieve==2.3.2.post1
urllib3==1.26.14

# Dev dependencies
pytest>=7.0
pytest-mock>=3.10
responses>=0.23
ruff>=0.1.0
pre-commit>=3.5.0
